Remove output scaling (#3)
Output scaling is not needed for wlock, as the entire buffer is set to a single color, and the current implementation causes an error on displays with scales >1. This change removes the unneeded listener function and scaling logic so that wlock works properly on displays that with a >1 scale factor.

Co-authored-by: Thanatos <thanatos@norelpy.codeberg.org>
Reviewed-on: https://codeberg.org/sewn/wlock/pulls/3
Co-authored-by: thanatos <thanatos@noreply.codeberg.org>
Co-committed-by: thanatos <thanatos@noreply.codeberg.org>